+++ src/Symfony/Component/Config/Loader/FileLoader.php
@@ -61,7 +61,7 @@ abstract class FileLoader extends Loader
      *
      * @param mixed       $resource       A Resource
      * @param string|null $type           The resource type or null if unknown
-     * @param bool        $ignoreErrors   Whether to ignore import errors or not
+     * @param int         $errorLevel     Whether to ignore import errors or not
      * @param string|null $sourceResource The original resource importing the new resource
      *
      * @return mixed
@@ -70,13 +70,23 @@ abstract class FileLoader extends Loader
      * @throws FileLoaderImportCircularReferenceException
      * @throws FileLocatorFileNotFoundException
      */
-    public function import($resource, $type = null, $ignoreErrors = false, $sourceResource = null)
+    public function import($resource, $type = null, $errorLevel = LoaderInterface::ERROR_LEVEL_ALL, $sourceResource = null)
     {
+        if (\is_bool($errorLevel)) {
+            @trigger_error('The boolean value you are using for the $errorLevel argument of the '.__CLASS__.'::import method is deprecated since Symfony 4.4 and will not be supported anymore in 5.0. Use the constants defined in the FileLoader instead.', E_USER_DEPRECATED);
+
+            if (true === $errorLevel) {
+                $errorLevel = LoaderInterface::ERROR_LEVEL_IGNORE_ALL;
+            } elseif (false === $errorLevel) {
+                $errorLevel = LoaderInterface::ERROR_LEVEL_ALL;
+            }
+        }
+
         if (\is_string($resource) && \strlen($resource) !== $i = strcspn($resource, '*?{[')) {
             $ret = [];
             $isSubpath = 0 !== $i && false !== strpos(substr($resource, 0, $i), '/');
-            foreach ($this->glob($resource, false, $_, $ignoreErrors || !$isSubpath) as $path => $info) {
-                if (null !== $res = $this->doImport($path, $type, $ignoreErrors, $sourceResource)) {
+            foreach ($this->glob($resource, false, $_, $errorLevel || !$isSubpath) as $path => $info) {
+                if (null !== $res = $this->doImport($path, $type, $errorLevel, $sourceResource)) {
                     $ret[] = $res;
                 }
                 $isSubpath = true;
@@ -87,13 +97,13 @@ abstract class FileLoader extends Loader
             }
         }
 
-        return $this->doImport($resource, $type, $ignoreErrors, $sourceResource);
+        return $this->doImport($resource, $type, $errorLevel, $sourceResource);
     }
 
     /**
      * @internal
      */
-    protected function glob(string $pattern, bool $recursive, &$resource = null, bool $ignoreErrors = false, bool $forExclusion = false, array $excluded = [])
+    protected function glob(string $pattern, bool $recursive, &$resource = null, int $errorLevel = LoaderInterface::ERROR_LEVEL_ALL, bool $forExclusion = false, array $excluded = [])
     {
         if (\strlen($pattern) === $i = strcspn($pattern, '*?{[')) {
             $prefix = $pattern;
@@ -109,7 +119,7 @@ abstract class FileLoader extends Loader
         try {
             $prefix = $this->locator->locate($prefix, $this->currentDir, true);
         } catch (FileLocatorFileNotFoundException $e) {
-            if (!$ignoreErrors) {
+            if ($errorLevel & LoaderInterface::ERROR_LEVEL_FILE_NOT_FOUND) {
                 throw $e;
             }
 
@@ -125,7 +135,7 @@ abstract class FileLoader extends Loader
         yield from $resource;
     }
 
-    private function doImport($resource, $type = null, bool $ignoreErrors = false, $sourceResource = null)
+    private function doImport($resource, $type = null, int $errorLevel = LoaderInterface::ERROR_LEVEL_ALL, $sourceResource = null)
     {
         try {
             $loader = $this->resolve($resource, $type);
@@ -157,7 +167,12 @@ abstract class FileLoader extends Loader
         } catch (FileLoaderImportCircularReferenceException $e) {
             throw $e;
         } catch (\Exception $e) {
-            if (!$ignoreErrors) {
+            if (LoaderInterface::ERROR_LEVEL_IGNORE_ALL !== $errorLevel) {
+                if (0 === ($errorLevel & LoaderInterface::ERROR_LEVEL_FILE_NOT_FOUND) && ($e instanceof FileLocatorFileNotFoundException)) {
+                    // ignore exception if file was not found
+                    return;
+                }
+
                 // prevent embedded imports from nesting multiple exceptions
                 if ($e instanceof LoaderLoadException) {
                     throw $e;